third_party/blink/web_tests/platform/generic/fast/parser/pre-first-line-break-expected.txt

Test for handling of line breaks following the pre element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS roundTrip('<pre></pre>') is '<pre></pre>'
PASS roundTrip('<pre>\n</pre>') is '<pre></pre>'
PASS roundTrip('<pre>\n\n</pre>') is '<pre>\n</pre>'
PASS roundTrip('<pre>x\n</pre>') is '<pre>x\n</pre>'
PASS roundTrip('<pre>&lt;\n</pre>') is '<pre>&lt;\n</pre>'
PASS roundTrip('<pre>&#61;\n</pre>') is '<pre>=\n</pre>'
PASS roundTrip('<pre><a></a></pre>') is '<pre><a></a></pre>'
PASS roundTrip('<pre><a>\n</a></pre>') is '<pre><a>\n</a></pre>'
PASS roundTrip('<pre>\n<a></a></pre>') is '<pre><a></a></pre>'
PASS roundTrip('<pre>\n<a>\n</a></pre>') is '<pre><a>\n</a></pre>'
PASS roundTrip('<listing></listing>') is '<listing></listing>'
PASS roundTrip('<listing>\n</listing>') is '<listing></listing>'
PASS roundTrip('<listing>\n\n</listing>') is '<listing>\n</listing>'
PASS roundTrip('<listing>x\n</listing>') is '<listing>x\n</listing>'
PASS roundTrip('<listing>&lt;\n</listing>') is '<listing>&lt;\n</listing>'
PASS roundTrip('<listing>&#61;\n</listing>') is '<listing>=\n</listing>'
PASS roundTrip('<listing><a></a></listing>') is '<listing><a></a></listing>'
PASS roundTrip('<listing><a>\n</a></listing>') is '<listing><a>\n</a></listing>'
PASS roundTrip('<listing>\n<a></a></listing>') is '<listing><a></a></listing>'
PASS roundTrip('<listing>\n<a>\n</a></listing>') is '<listing><a>\n</a></listing>'
PASS roundTrip('<div></div>') is '<div></div>'
PASS roundTrip('<div>\n</div>') is '<div>\n</div>'
PASS roundTrip('<div>\n\n</div>') is '<div>\n\n</div>'
PASS roundTrip('<div>x\n</div>') is '<div>x\n</div>'
PASS roundTrip('<div>&lt;\n</div>') is '<div>&lt;\n</div>'
PASS roundTrip('<div>&#61;\n</div>') is '<div>=\n</div>'
PASS roundTrip('<div><a></a></div>') is '<div><a></a></div>'
PASS roundTrip('<div><a>\n</a></div>') is '<div><a>\n</a></div>'
PASS roundTrip('<div>\n<a></a></div>') is '<div>\n<a></a></div>'
PASS roundTrip('<div>\n<a>\n</a></div>') is '<div>\n<a>\n</a></div>'
PASS successfullyParsed is true

TEST COMPLETE
